fix: cmd/agent переименовать файлы для точности
